Refactoring the UDTValueReader and Writer.
