+++ spring-data-cassandra/src/main/java/org/springframework/data/cassandra/convert/UDTValueWriter.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2013-2014 the original author or authors
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.springframework.data.cassandra.convert;
 
 import java.util.ArrayList;
@@ -8,18 +23,18 @@ import com.datastax.driver.core.UDTValue;
 import com.datastax.driver.core.UserType;
 
 /**
- * Helpful class to read a column's value from a row, with possible type conversion.
+ * Helpful class to write a user defined type value from a row, with possible type conversion.
  * 
- * @author Matthew T. Adams
+ * @author Fabio Mendes
  */
-public class UDTValueAccessor extends AbstractColumnAccessor {
-
+public class UDTValueWriter extends AbstractColumnWriter {
+	
 	protected UDTValue udtValue;
 	protected UserType udtType;
 	private List<String> fieldNames;
 
-	public UDTValueAccessor(UDTValue value) {
-		super(value, value);
+	public UDTValueWriter(UDTValue value) {
+		super(value);
 		this.udtValue = value;
 		this.udtType = value.getType();
 		this.fieldNames = new ArrayList<String>(udtType.getFieldNames());
@@ -41,4 +56,5 @@ public class UDTValueAccessor extends AbstractColumnAccessor {
 	public UDTValue getUDTValue() {
 		return udtValue;
 	}
+
 }
